requirements.txt
anytree==2.8.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.13

# 4.1.2
django-admin-csvexport==1.11
django-allauth==0.51.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-import-export==2.9.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
types-cryptography==3.3.23
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
Babel==2.10.3


# phone-number dpendencies
django-phonenumber-field==7.0.0
django-label-templates==1.0.3
phonenumbers==8.12.57

# django-phonenumbers==1.0.